renamed:    ZXCPMsrc/FatFs/diskio.c -> GTKlinux/Src/FatFs/diskio.c 	renamed:    ZXCPMsrc/FatFs/diskio.h -> GTKlinux/Src/FatFs/diskio.h 	renamed:    ZXCPMsrc/FatFs/fatfs.c -> GTKlinux/Src/FatFs/fatfs.c 	renamed:    ZXCPMsrc/FatFs/fatfs.h -> GTKlinux/Src/FatFs/fatfs.h 	renamed:    ZXCPMsrc/FatFs/ff.c -> GTKlinux/Src/FatFs/ff.c 	renamed:    ZXCPMsrc/FatFs/ff.h -> GTKlinux/Src/FatFs/ff.h 	renamed:    ZXCPMsrc/FatFs/ff_gen_drv.c -> GTKlinux/Src/FatFs/ff_gen_drv.c 	renamed:    ZXCPMsrc/FatFs/ff_gen_drv.h -> GTKlinux/Src/FatFs/ff_gen_drv.h 	renamed:    ZXCPMsrc/FatFs/ffconf.h -> GTKlinux/Src/FatFs/ffconf.h 	renamed:    ZXCPMsrc/FatFs/integer.h -> GTKlinux/Src/FatFs/integer.h 	modified:   STM32CubeIDE/.project 	new file:   STM32CubeIDE/Inc/fatfs.h 	new file:   STM32CubeIDE/Inc/ffconf.h 	modified:   STM32CubeIDE/Inc/main.h 	modified:   STM32CubeIDE/Inc/stm32_assert.h 	new file:   STM32CubeIDE/Inc/stm32f1xx_hal_conf.h 	renamed:    ZXCPMsrc/user_diskio.h -> STM32CubeIDE/Inc/user_diskio.h 	new file:   STM32CubeIDE/Src/fatfs.c 	modified:   STM32CubeIDE/Src/main.c 	new file:   STM32CubeIDE/Src/stm32f1xx_hal_msp.c 	modified:   STM32CubeIDE/Src/stm32f1xx_it.c 	renamed:    ZXCPMsrc/user_diskio.c -> STM32CubeIDE/Src/user_diskio.c 	modified:   STM32CubeIDE/ZX_and_CPM_on_STM32.ioc 	modified:   ZXCPMsrc/SD_driver.c 	modified:   ZXCPMsrc/SD_driver.h 	modified:   ZXCPMsrc/z80.c 	modified:   ZXCPMsrc/z80_loader.c

+++ ZXCPMsrc/z80.c
@@ -63,7 +63,7 @@ void z80_interrupt() {
 		break;
 	case 2:
 	{
-		register uint16_t inttemp=(0x100*I)+0xff;
+		register uint32_t inttemp=(0x100*I)+0xff;
 		PCL = mem_read(inttemp++); PCH = mem_read(inttemp);
 		z80_tstates += 6;
 		break;
